+++ ui/packages/consul-ui/app/components/hashicorp-consul/index.hbs
@@ -136,14 +136,9 @@
           <a href={{href-to 'dc.intentions' @dc.Name params=(hash peer=undefined)}}>Intentions</a>
         </li>
       {{/if}}
-      {{#if (can 'read intentions')}}
-        <li
-          data-test-main-nav-application-map
-          class={{if (is-href 'dc.intentions' @dc.Name) 'is-active'}}
-        >
-          <a href={{href-to 'dc.intentions' @dc.Name params=(hash peer=undefined)}}>Application Map</a>
-        </li>
-      {{/if}}
+      <li data-test-main-nav-api-endpoints class={{if (is-href 'dc.poapis' @dc.Name) 'is-active'}}>
+        <a href={{href-to 'dc.poapis' @dc.Name params=(hash peer=undefined)}}>API Endpoints</a>
+      </li>
       <Consul::Acl::Selector @dc={{@dc}} @partition={{@partition}} @nspace={{@nspace}} />
       <Consul::Peer::Selector @dc={{@dc}} @partition={{@partition}} @nspace={{@nspace}} />
     </ul>
