+++ ui/packages/consul-ui/app/styles/app.scss
@@ -97,3 +97,21 @@
   height: 1.2rem;
   width: auto;
 }
+
+.po-container {
+  width: 100%;
+  height: 100%;
+}
+
+.display-flex-container {
+  display: flex;
+  margin: 1rem 0rem;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.api-endpoints-title {
+  white-space: nowrap;
+  font-size: 1.5rem;
+  font-weight: bold;
+}
